refactor: :truck: Organizei melhor os schemas de banco.
